requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-pptx==0.6.21
pillow==10.1.0
requests==2.31.0
openai==1.6.1
anthropic==0.7.7
google-generativeai==0.3.0
werkzeug==3.0.1
httpx==0.25.0
httpcore==0.18.0
h11==0.14.0

